Create WDCGAN_for_Consumption_Data.ipynb
For the privacy of electricity users' personal electricity data, a large amount of highly fine-grained electricity data cannot be shared and studied by researchers, which restricts the use and validation of various deep models in the electricity system. For this reason, we use WDCGAN to generate time-series electricity data of users and use the generated data instead of real data for research and use, and this approach can find a balance between data availability and privacy balance.